core: move header `LoadCoalGeometries.h` and its functions to library core
