+++ src/candlewick/core/LoadCoalGeometries.cpp
