+++ src/candlewick/core/LoadCoalGeometries.h
